eme_playback/test_eme_playback.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Encrypted Media Extensions</title>
  <script type="text/javascript" src="SimpleTest.js"></script>
  <script type="text/javascript" src="manifest.js"></script>
  <script type="text/javascript" src="eme.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

function startTest(test, token)
{
  var sessions = [];
  function onSessionCreated(session) {
    sessions.push(session);
  }

  let v = document.createElement("video");
  document.body.appendChild(v);

  var gotEncrypted = 0;
  let finish = new EMEPromise;

  v.addEventListener("ended", function(ev) {
    ok(true, TimeStamp(token) + " got ended event");

  });

  Promise.all([
    LoadInitData(v, test, token),
    CreateAndSetMediaKeys(v, test, token),
    LoadTest(test, v, token)])
  .then(values => {
    console.log(" Start to play ENCRYPTED content !!");
    v.play();
    let initData = values[0];
    initData.map(ev => {
      let session = v.mediaKeys.createSession();
      onSessionCreated(session);
      MakeRequest(test, token, ev, session);
    });
  })
  .then(() => {
    console.log(" finish.promise ");
    return finish.promise;
  })
  .catch(reason => ok(false, reason))
}
startTest(gEMETests[0], 'test>>>>');
</script>
</pre>
</body>
</html>
